+++ R/paste.R
@@ -10,8 +10,7 @@
 #' # get function from namespace instead of possibly getting
 #' # implementation shipped with recent R versions:
 #' bp_paste = getFromNamespace("paste", "backports")
-#' x = matrix(1:6, 2, 3)
-#' bp_paste(x, 1)
+#' bp_paste(letters[1:3], character(), collapse = NULL, recycle0 = TRUE)
 paste <- function(..., sep = " ", collapse = NULL, recycle0 = FALSE) {
   if(recycle0 && any(vapply(list(...), length, FUN.VALUE = NA_integer_) == 0L)) {
     base::paste(sep = sep, collapse = collapse) # for argument checking